+++ src/giskard_hub/data/evaluation.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import math
 import time
-import warnings
 from dataclasses import dataclass, field
 from time import sleep
 from typing import Any, Dict, List
